Basic QR code generator/deep link app.
When accessed from the web with no parameters, it displays the generator.
Scan the code with your mobile and it will open the Instagram app for the username you provided.
If the app is not installed, it will automatically redirect you to App Store or Play Store, depending on your device.